wip: removed unused tests
